moved embedding utils out of common/utils.py
