+++ align_data/finetuning/finetuning_dataset.py
@@ -9,7 +9,7 @@ from sqlalchemy.exc import OperationalError
 from sqlalchemy.sql import func
 
 from align_data.db.session import make_session
-from align_data.common.utils import get_embedding
+from align_data.pinecone.embedding_utils import get_embedding
 from align_data.pinecone.pinecone_db_handler import PineconeDB
 from align_data.pinecone.text_splitter import ParagraphSentenceUnitTextSplitter
 from align_data.pinecone.update_pinecone import get_text_chunks
