+++ align_data/pinecone/embedding_utils.py
